+++ src/pages/CubePlaytestPage.js
@@ -30,7 +30,7 @@ import CubeContext from 'components/CubeContext';
 import CustomDraftFormatModal from 'components/CustomDraftFormatModal';
 import DynamicFlash from 'components/DynamicFlash';
 import DeckPreview from 'components/DeckPreview';
-import Markdown from 'components/MarkdownRenderer';
+import Markdown from 'components/Markdown';
 import withModal from 'components/WithModal';
 import useAlerts, { Alerts } from 'hooks/UseAlerts';
 import useToggle from 'hooks/UseToggle';
