+++ src/pages/MarkdownPage.js
@@ -7,7 +7,7 @@ import DynamicFlash from 'components/DynamicFlash';
 import Advertisement from 'components/Advertisement';
 import MainLayout from 'layouts/MainLayout';
 import RenderToRoot from 'utils/RenderToRoot';
-import Markdown from 'components/MarkdownRenderer';
+import Markdown from 'components/Markdown';
 
 const MarkdownPage = ({ user, loginCallback }) => (
   <MainLayout loginCallback={loginCallback} user={user}>
